+++ test/core-filters/setupFilters-skipIfSmaller.test.js
@@ -1,14 +1,14 @@
 const corefilters = require("../../lib/core-filters");
 
-test("filter/setupFilters/skipIfNewer/true", () => {
+test("filter/setupFilters/skipIfSmaller/true", () => {
     var workingObject = {
         validatedOption: {
             skipIfExists: false,
             skipIfNotExists: false,
-            skipIfNewer: true,
+            skipIfNewer: false,
             skipIfOlder: false,
             skipIfBigger: false,
-            skipIfSmaller: false
+            skipIfSmaller: true
         },
         fileFilters: []
     };
@@ -16,11 +16,11 @@ test("filter/setupFilters/skipIfNewer/true", () => {
     return expect(corefilters.setupFilters(workingObject))
         .resolves
         .toMatchObject({
-            fileFilters: [corefilters.skipIfNewer]
+            fileFilters: [corefilters.skipIfSmaller]
         });
 });
 
-test("filter/setupFilters/skipIfNewer/false", () => {
+test("filter/setupFilters/skipIfSmaller/false", () => {
     var workingObject = {
         validatedOption: {
             skipIfExists: true,
@@ -36,6 +36,6 @@ test("filter/setupFilters/skipIfNewer/false", () => {
     return expect(corefilters.setupFilters(workingObject))
         .resolves.not
         .toMatchObject({
-            fileFilters: [corefilters.skipIfNewer]
+            fileFilters: [corefilters.skipIfSmaller]
         });
 });